+++ metrics/metrics-micrometer-core/src/main/java/uk/gov/justice/services/metrics/micrometer/bootstrap/CounterMeterRegistrar.java
@@ -1,0 +1,16 @@
+package uk.gov.justice.services.metrics.micrometer.bootstrap;
+
+import uk.gov.justice.services.metrics.micrometer.meters.MetricsMeter;
+
+import io.micrometer.core.instrument.Counter;
+import io.micrometer.core.instrument.MeterRegistry;
+
+public class CounterMeterRegistrar {
+
+    public void registerCounterMeter(final MetricsMeter metricsMeter, final MeterRegistry meterRegistry) {
+        Counter.builder(metricsMeter.metricName())
+                .tags(metricsMeter.metricTags())
+                .description(metricsMeter.metricDescription())
+                .register(meterRegistry);
+    }
+}