+++ metrics/metrics-micrometer-core/src/main/java/uk/gov/justice/services/metrics/micrometer/config/MicometerMetricsConfigurationException.java
@@ -1,0 +1,8 @@
+package uk.gov.justice.services.metrics.micrometer.config;
+
+public class MicometerMetricsConfigurationException extends RuntimeException {
+
+    public MicometerMetricsConfigurationException(final String message) {
+        super(message);
+    }
+}