+++ metrics/metrics-micrometer-core/src/main/java/uk/gov/justice/services/metrics/micrometer/meters/MetricsProviderException.java
@@ -1,0 +1,8 @@
+package uk.gov.justice.services.metrics.micrometer.meters;
+
+public class MetricsProviderException extends RuntimeException {
+
+    public MetricsProviderException(final String message, final Throwable cause) {
+        super(message, cause);
+    }
+}