+++ metrics/metrics-micrometer-core/src/main/java/uk/gov/justice/services/metrics/micrometer/registry/MetricsRegistrar.java
@@ -1,0 +1,11 @@
+package uk.gov.justice.services.metrics.micrometer.registry;
+
+import io.micrometer.core.instrument.MeterRegistry;
+import io.micrometer.core.instrument.Metrics;
+
+public class MetricsRegistrar {
+
+    public void addRegistry(final MeterRegistry meterRegistry) {
+        Metrics.addRegistry(meterRegistry);
+    }
+}